refactor: extract EvalCallback, gate diagnostic metrics behind flag
- Move _EvalCallback from train.py into scripts/rsl_rl/eval_callback.py
- Add ENABLE_ADDITIONAL_METRICS = False flag in hand_object_commands.py;
  all diagnostic metrics added on this branch (contact wrench, bbox corner
  error, contact fracs, hand tracking quality, voc_decay_fired) are gated
  behind it in both _init_metrics and _update_metrics
- Reduce articulated object joint friction 1.0 → 0.1 in apply_scene_config
